+++ backend/src/main/java/washit/backend/repository/WashingMachineRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import washit.backend.model.WashingMachine;
 
 @Repository
-public interface WashingMachineRepository extends JpaRepository<WashingMachine, Long> {
+public interface WashingMachineRepository extends JpaRepository<WashingMachine, Integer> {
 
 }
